refactor server code, making game rooms for groups of two players
