+++ server/package.json
@@ -11,7 +11,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.20.0",
     "express": "^4.17.3",
     "mongodb": "^4.5.0",
     "mongoose": "^6.3.0",
