+++ server/server.js
@@ -4,9 +4,6 @@ const morgan = require("morgan"); // Server logs
 const dotenv = require("dotenv"); // Load ENV
 const mongoose = require("mongoose");
 const bp = require("body-parser");
-const { createServer } = require("http");
-const { Server: SocketServer } = require("socket.io");
-const PORT = process.env.PORT || 3001;
 
 require("dotenv").config();
 
@@ -23,35 +20,14 @@ db.on("error", console.error.bind(console, "MongoDB connection error:"));
 // Middleware
 app.use(morgan("dev"));
 app.use(express.static("public")); // Serve static files
-app.use(bp.json());
-app.use(bp.urlencoded({ extended: true }));
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
 dotenv.config();
-
-// SocketIO
-const httpServer = createServer(app);
-const io = new SocketServer(httpServer, {
-  cors: {
-    origin: '*'
-  },
-});
-
-io.on("connection", (socket) => {
-  console.log(`connection created for user ${socket.id}`);
-  socket.emit("newServerMessage", { body: "test server message" });
-
-  socket.on("newClientMessage", (arg) => {
-    console.log(`user ${socket.id} replied with ${arg}`);
-  });
-
-  socket.on('progressUpdate', (arg) => {
-    console.log(`user ${socket.id} progressed and ${arg}`)
-  });
-});
 
 // Initialize Routes
 const userRoutes = require("./routes/user");
 const promptRoutes = require("./routes/prompt");
-const bodyParser = require("body-parser");
+const router = require("./routes/user");
 
 app.use("/", userRoutes);
 app.use("/prompts", promptRoutes);
@@ -60,6 +36,4 @@ app.get("/home", (req, res) => {
   res.send("Hello|World");
 });
 
-httpServer.listen(PORT, () => {
-  console.log(`http server listening on port ${PORT}`);
-});
+module.exports = app;
